requirements.txt
certifi
pytz
setuptools
attrs
pip
Faker~=15.2.0
qrcode
multidict
gspread~=5.6.2
cachetools
protobuf
rsa
oauth2client~=4.1.3
uritemplate
async-timeout
aiohttp
idna
oauthlib
openpyxl
pyparsing
requests~=2.28.1
google-auth
python-dateutil
charset-normalizer
urllib3
numpy
six
yarl
pandas~=1.5.1
frozenlist
requests-oauthlib
PyQRCode~=1.2.1
aiosignal
et-xmlfile
pypng
wheel
httplib2
google-auth-oauthlib
pyasn1
colorama
pyasn1-modules
imgbbpy~=0.1.3
